third_party/WebKit/LayoutTests/external/wpt/webusb/worker/usb.serviceworker.https-expected.txt

This is a testharness.js-based test.
PASS Ensure that WebUSB is inaccessible from a service worker.
FAIL Service workers should not have access to the WebUSB API. assert_equals: navigator.usb should not be a USB object expected "undefined" but got "object"
Harness: the test ran to completion.
